Use minified JS package (#1)
This is my attempt at integrating the data source we generated into the emoji-mart package. It reduces the size of emoji-mart.min.js from 763KB to 177KB and is faster to initialize. I also cut out a lot of the extra stuff from the library that we don't want it to do (different datasets, emoticons, etc).

However, this comes with a tradeoff. The file we generated is missing a lot of the information the original data source has, which means we can't search by "metadata", just matches on short names. I'm not sure how big a problem this is - the search for "smile" is much smaller now but most searches are the same.