remove eventlistener that is never removed (#13)
Removes an added EventListener that is never removed. This has also been reported to the upstream repo as issue missive/emoji-mart#635

Unfortunately, I tried to simply remove it with an componentWillUnmount, but the function was never called. I tried to debug why, but couldn't figure it out (noticeable, preact has some reported problems with componnetWillUnmount). In any case, handleClickOutside is unused by Figma. Secondly, I think the only time we would use it is to close the emoji picker, and it is not very react-like for components to close themselves.

I'm open to other suggestions / pairing on this issue, but I have doubts that the time spent will be worth it.