Render native emojis in picker if 'set' is 'native' (#18)
While working on this PR, I found that our forked version of emoji-mart always renders images of emojis regardless of the set property of the Picker. This change modifies the Emoji component to render native emojis if the set property is native. This will allow us to have a native emoji picker in comments, but still allow the image-based emoji picker for editor text.